feat: history에서 식습관 클릭시 맵으로 이동하여 그 주변 식당 체킹
