+++ apps/client/components/Map/Button/RecentButton.tsx
@@ -3,9 +3,22 @@ import { Recent } from "@svgs/map";
 import { flex, position } from "@toss/emotion-utils";
 import { motion } from "framer-motion";
 import { defaultScaleChangeVariants, framerMocker } from "~/constants";
+import { useMapBoundsValue, useMapSearchByCountrySetter } from "~/recoil/atoms";
+import { useFetchDefaultMap } from "~/server/recent";
 
 const RecentButton = () => {
   const { colors, zIndex } = useTheme();
+  const setMapSearchByCountry = useMapSearchByCountrySetter();
+  const mapBounds = useMapBoundsValue();
+  const { mutate: defaultMap } = useFetchDefaultMap({
+    onSuccess: data => {
+      setMapSearchByCountry(data);
+    }
+  });
+
+  const clickHandler = () => {
+    if (mapBounds) defaultMap(mapBounds);
+  };
 
   return (
     <motion.button
@@ -13,6 +26,7 @@ const RecentButton = () => {
       {...framerMocker}
       whileHover="whileHover"
       variants={defaultScaleChangeVariants}
+      onClick={clickHandler}
       css={css`
         ${position("absolute", {
           top: 52,
